Fix regex for concrete subclass filter
Swift enforces subclasses must be the first item in the inheritance list. Therefore, the regex does not need to include characters before the abstract class name.

Also removed unnecessary field parsing of return type and parameter types, since they are not, and cannot be used in validation due to generics.